refactor to use pydantic all through the procedure state across actions and calibration data
