+++ evolver/calibration/procedure.py
@@ -1,32 +1,10 @@
 from abc import ABC
-from copy import deepcopy
-from typing import Any, Dict, List
+from typing import Any, Dict
 
-from pydantic import BaseModel, Field
 
 from evolver.base import BaseInterface
 from evolver.calibration.action import CalibrationAction
-
-
-class CalibrationStateModel(BaseModel):
-    """
-    Model to represent the state of a calibration procedure. All procedures record their completed actions, and history of actions in this model.
-    The data collected by the calibration procedure (i.e. the data the actions have gathered, that's used as input to the Hardware.input/outputTransformer methods) is also stored here.
-
-    Attributes:
-        started (bool): A flag to indicate if the calibration procedure has been initialized, used by the front end to determine procedure controls to display.
-        completed_actions (List[str]): A list of actions that have been completed during the calibration procedure.
-        history: A list of previous states of the calibration procedure. Used to undo actions.
-        measured (Dict[str, Any]): A dictionary of data collected by the calibration procedure. This data is used by the Transformer class to fit a model to the data. For example, a temperature calibrator might collect raw and reference temperature data for each vial.
-    """
-
-    class Config:
-        extra = "allow"
-
-    completed_actions: List[str] = Field(default_factory=list)
-    history: List["CalibrationStateModel"] = Field(default_factory=list)
-    started: bool = False
-    measured: Dict[Any, Any] = {}
+from evolver.calibration.state import CalibrationStateModel
 
 
 class CalibrationProcedure(BaseInterface, ABC):
@@ -56,8 +34,8 @@ class CalibrationProcedure(BaseInterface, ABC):
         """
         super().__init__(*args, **kwargs)
         self.actions = []
-        self.state = CalibrationStateModel(**(state or {})).model_dump()
-        self.state["started"] = True
+        self.state = CalibrationStateModel(**(state or {}))
+        self.state.started = True
         self.hardware = hardware
 
     def add_action(self, action: CalibrationAction):
@@ -78,8 +56,8 @@ class CalibrationProcedure(BaseInterface, ABC):
         """
         Undo the last action that was dispatched in the calibration procedure.
         """
-        if len(self.state["history"]) > 0:
-            self.state = self.state["history"].pop()
+        if len(self.state.history) > 0:
+            self.state = self.state.history.pop()
         return self.state
 
     def save(self):
@@ -92,16 +70,17 @@ class CalibrationProcedure(BaseInterface, ABC):
         # calibration_file maybe none, in which case the save operation must fail with an error message.
         if file_path is None:
             raise ValueError("calibration_file attribute is not set on the Calibrator config.")
-        self.hardware.calibrator.calibration_data.procedure_state = {**self.state}
+        self.hardware.calibrator.calibration_data.procedure_state = self.state
         self.hardware.calibrator.calibration_data.save(file_path)
         return self.state
 
     def dispatch(self, action: CalibrationAction, payload: Dict[str, Any]):
         if payload is not None and action.FormModel.model_fields != {}:
             payload = action.FormModel(**payload)
-        previous_state = deepcopy(self.state)
+        previous_state = self.state.model_dump()
         updated_state = action.execute(self.state, payload)
-        updated_state["completed_actions"].append(action.name)
-        updated_state["history"].append(previous_state)
+        updated_state.completed_actions.append(action.name)
+        # Convert previous state to model before appending
+        updated_state.history.append(CalibrationStateModel.model_validate(previous_state))
         self.state = updated_state
         return self.state
